+++ CHANGELOG.md
@@ -10,6 +10,7 @@ increment the `minor` version for breaking changes.
 ### Added
 - config macro for win32 `<memoryapi.h>` Virtual functions
 - config macro for posix `<sys/mman.h>` `m(un)map` and `mprotect` functions
+- config macro for `__attribute__((sysv_abi))`
 - config macro for x86 target (enabled on 16bit, 32bit, and 64bit)
 - config macro for 64bit x86 target
 - `osmem.h` in TSX impl, providing `alloc_rw`, `protect_rx`, and `free`